Error handling improvements in PKI CLI.
The MainCLI has been modified to improve the error handling.
When a generic Exception is caught, the exception will be
displayed if it contains an error message, otherwise it will
exit silently. When a ProcessingException is caught, the
cause of the exception will be displayed if it's available,
otherwise the exception itself will be displayed.