Fixed exception handling in PKIConnection
When an error occurs on the server, it will return a response
containing the exception info to the client, and the client
is supposed to recreate and rethrow the exception on the client
side.

Previously the client would use MediaType.equals() to check
the content type of the response. If the content type is an
application/xml or an application/json, the client can parse
the exception info for recreating the exception.

However, since the actual content type contains a charset
parameter (e.g. application/xml;charset=utf-8), the code could
not match it against any of the supported types, so it threw a
generic PKIException instead.

Now the code has been modified to use MediaType.isCompatible()
which will match the content type properly regardless of the
charset parameter, so the client can throw the proper exception.

https://bugzilla.redhat.com/show_bug.cgi?id=1778953