Fixed failures in pkidestroy
Previously pkidestroy would shutdown and restart the
server when removing a subsystem from an instance.
Usually it takes sometime before the server becomes
ready again. If multiple pkidestroy operations are
executed consecutively (as in the CI), some of the
operations might fail since the server is not ready
yet due to the previous restart.

Instead of restarting the server, the uninstall
scriptlets have been modified to only undeploy the
subsystem web application while keeping the server
running. When the last subsystem is removed, the
server will finally be shutdown.

https://pagure.io/dogtagpki/issue/2804