Fix support for SoftHSM
The NSSDatabase.add_cert() has been modified to accept
an optional user param to be used for running certutil
when importing a cert.

pkispawn has been modified to import the permanent SSL
cert using the instance user. This way if the instance
is configured to use SoftHSM, the files will be created
with the right permission.

Resolves: https://github.com/dogtagpki/pki/issues/3204