Fixed exception handling in LDAPConfigurator.createReplicaObject()
Originally the LDAPConfigurator.createReplicaObject() would
return true if it managed to add a new replica object. If the
object already existed, it would only add the new bind DN and
return false. If an error happened it would get ignored and
the method would return false as well.

In 4abfdc77508545fb90ef127fbbf373ae1609d705 the behavior of
accidentally got changed during clean up to return true if the
replica object already exists and this caused OCSP clone of
clone installation to fail.

To fix the problem the behavior has been reverted except that
now any error will be reported as an exception.