Fix pkidestroy for KRA with external certs
Normally when installing KRA the CA will also create a user
which later can be used by pkidestroy to remove the KRA
connector from the CA. However, if the KRA was installed with
external certs the user would not be created so pkidestroy
could not remove the connector.

To fix the problem pkispawn has been modified to add a user
into the CA with the required authorization if it does not
exist already so it can be used later by pkidestroy.