Fix pkidestroy for KRA with external certs
Normally when installing KRA a KRA connector will be added into the
CA and a user will be created in the CA for managing the connector.
In case the KRA is removed later pkidestroy can use this user to
remove the KRA connector from the CA.

However, if the KRA was installed with external certs (including
CMC case) the user would not be created so pkidestroy cannot remove
the KRA connector while removing the KRA.

To fix the problem some of the code in SubsystemGroupUpdater has
been copied into UpdateConnector such that the CA will create the
user while adding the KRA connector in all cases. In the future
the SubsystemGroupUpdater itself might be removed.