Fix KeyId to preserve leading zeros
Previously KeyId was implemented using a BigInteger which
does not preserve leading zeros. This could cause a problem
when searching for a key in NSS database by its ID.

To fix the problem the KeyId has been modified to store the
ID as a byte array so it will preserve the original ID.

The KeyIdTest has been updated to test additional scenarios.