Add single-stage CA Dockerfile
The multi-stage Dockerfile at the main source folder is used
mainly for CI. This Dockerfile requires that the binaries be
be built prior to building the container.

In order to build on Quay, the Dockerfile must build the
binaries within the container itself using a single stage,
so a new Dockerfile has been added for CA for this purpose.
For efficiency, this Dockerfile will build plain binaries
without the RPM packages.