Add cache for pki-deps image
The CI build jobs have been modified to build pki-deps image
then store it in GH cache. In the next CI execution the image
will be loaded from the cache, so it could shorten the build
time by about 3 minutes. The cache will be invalidated if the
pki.spec (which defines the dependencies) is modified.