Refactor CMSGateway
Previously the CMSGateway used the static CMS.getCMSEngine()
which could potentially cause class loading issues. To avoid
the problem the static fields and methods in the class have
been converted into non-static, and a new engine field has
been added to store a reference to the CMSEngine instance
such that it no longer needs to rely on CMS.getCMSEngine().