Fix race condition during ACME order finalization
Previously when the authorization for an order completed the
ACMEChallengeProcessor would update the authorization status
and the order status as well, then the client would call the
ACMEFinalizeOrderService to finalize the order.

However, since these updates are not atomic there is a risk
that the client will detect the new authorization status then
immediately call the ACMEFinalizeOrderService before the order
status can be updated by ACMEChallengeProcessor. Since both
both ACMEFinalizeOrderService and ACMEChallengeProcessor are
trying to read/write the same order at the same time, the
finalization process could fail.

To avoid the problem the ACMEChallengeProcessor has been
modified to update the order status first before updating the
authorization status.