Update Azure pipelines to support BASE_IMAGE
The Azure pipelines has been modified to support BASE_IMAGE param
so it can be used to test different platforms.

The pki.spec has been modified to use maven-local-openjdk<version>
to ensure that the code is built with the Java version currently
supported (i.e. OpenJDK 17) regardless of the platform.

In the future it might be possible to configure the version number
to test different Java versions.

https://github.com/dogtagpki/pki/wiki/Configuring-Test-OS