Update test scripts
The scripts for creating, starting, and stopping test containers
have been modified to provide an option to specify the image to
use. By default it will use pki-base image which is basically a
Fedora image with systemd enabled. For now the CI will continue
to use the DS image from quay.io/389ds/dirsrv in most tests.