Add pki --api option
The pki CLI has been updated to provide an option to specify
the REST API version to use when communicating with the server.
Currently the default is set to v1, but it might change later.

The PKIClient class has been modified to store the API version
which will automatically be used by other client classes (e.g.
InfoClient).

The pki info has been modified to show the API version in use.
The basic CA test has been updated to run pki info with the
option and verify the access logs generated by these commands.