Add installation params for range DNs
The pki_<object>_number_range_dn params have been added to
provide a way to customize the subtrees used to store the
range objects for certs and requests in CA.

The SubsystemDBInitCLI has been modified to initialize the
database based on the range DN configuration. The hard-coded
subtrees in the create.ldif have been removed.

The Repository.getNextRangeDN() has been added to return the
DN of the entry holding the nextRange attribute. The methods
in SubsystemRangeUpdateCLI have also been modified to take
the DN of the entry holding the nextRange attribute. These
changes will make it easier to change the location of the
nextRange attribute in the future.