Add installation params for range subtrees
The pki_<object>_number_range_dn params have been added to
provide a way to customize the subtrees used to store the
range objects for certs and requests in CA.

The Repository.getNextRangeDN() has been added to return the
DN of the entry that stores the nextRange attribute. This
will make it easier to change in the future.

The SubsystemDBInitCLI has been modified to create the range
subtrees based on the configuration. The hard-coded subtrees
in the create.ldif have been removed.

The methods in SubsystemRangeUpdateCLI have been modified to
take the DN of the entry that holds the nextRange attribute
to make it easier to change in the future.