Fix ca_admin_cert.p12
Normally when installing a CA the admin cert will be stored in
~/.dogtag/<instance>/ca_admin_cert.p12. If the CA is removed
the file is not automatically removed since it's owned by the
admin user and might be used again in the future.

Previously, if the CA was reinstalled with a new CA signing cert
pkispawn would import the ca_admin_cert.p12 if it exists, but the
admin cert would be unusable since it's issued by the old CA.

To address the problem pkispawn has been modified to validate
the provided admin cert and halt the installation if the cert is
invalid (e.g. issued by unknown CA). The admin user will be
responsible to remove the invalid cert then run pkispawn again.

The test for installing CA with existing config has been updated
to reinstall the CA with new system certs but with an old admin
cert in ca_admin_cert.p12 (which should fail), remove the old
admin cert, then reinstall the CA again so it will generate a
new admin cert (which should work).

Resolves: https://github.com/dogtagpki/pki/issues/4936