Don't create audit signing cert by default
Previously the audit signing cert was always created by default,
but it's actually only used if log signing is enabled.

To reduce redundancies the default pki_audit_signing_nickname has
been removed, and pkispawn has been modified to skip creating the
audit signing cert if the nickname is not specified.

The selftest has also been modified to skip validating the audit
signing cert if the nickname is missing.

The pki-server cert-find and pki-server subsystem-cert-find
commands have been modified to skip displaying the audit signing
cert if the nickname is missing.

The basic tests of each subsystem have been updated to install
the subsystem without the audit signing cert nickname so it won't
create the cert. For basic CA test, later it will create the cert
manually and enable log signing.