Refactor AuthorityRepository.getCA()
The AuthorityRepository.getCA() has been modified to get the
authority data from the authority record directly. For now it
still uses the CA object to get some additional info, but in
the future it might be possible to store those info in LDAP as
well.

The AuthorityService.getCA() has been modified to use the
AuthorityRepository.getCA().