Update standalone KRA test
The test for standalone KRA has been updated to install the KRA
without a security domain and while the CA is down so the KRA
connector in CA and the CA subsystem user in KRA need to be set
up separately after installation.

In the future all KRA installations might be done this way to
reduce the complexity of the installation process and to make
it more reliable (i.e. not dependent on a running CA).

The pki-server status has been updated to support KRA that does
not have any security domain params.