Add i18n test
The PKIClient has been updated to encode the request text into
UTF-8 before sending it to the server.

The pki <subsystem>-user-add has been modified to accept a
--full-name option for consistency with the pki-server
<subsystem>-user-add command.

The Python AddUserCLI has been modified to read the proper
fields from the JSON response.

A new test has been added to create a custom profile with
wide characters in the name and description, enroll a cert
with it, create a system user with wide characters in the
full name, assign the cert to the user, and auth with the
cert.